Update index.html
for private deployment, please see blank.json as a good starting point for unbranded customization text.  My goal is to move towards more of a library for the core code, so I am trying to separate out the UI/text/style more and more. This should make forking of new branches of the code easier in the long run.