v10.2 - bug fixes
Fixed a bug with the group room, where videos froze on guest re-join

added &scene=0 (&scene) , which auto-adds videos to the scene without interaction

Updated the &mirror URL flag to have 0, 1, 2, and 3 as options.

-no mirror parameter in the URL is the default.  Local previews are mirrored
-mirror=0 turns off ALL mirroring, for everything.
-mirror=1 inverts the default.  Local previews are not mirrored, but guests are mirrored.
-mirror=2 mirrors ALL videos
-mirror=3 is the same as default, except the entire website is mirrored, including Text (useful for teleprompters?)

Experimental Code Change :  When the data-channel closes, I auto-close the video channels as well, and try to reload the video from scratch.  I do not wait for the video stream to time out.  This *might* fix a 'stacking video' problem, as well as reducing the "frozen frame" video problem when a guest reloads.

This 10.2 release is currently on https://obs.ninja/beta for testing