Add a dedicated Refresh button
The original onclick="history.go(0);" method only forces the browser to reload the page from the cache. But if we use iframes, that will not work as expected if we try/debug different parameters in URL(iframed). This dedicated button uses location.reload() method, which will reload the current URL in the parent window of the current frame. This will cause the entire parent window, including any frames or iframes, to reload, reflecting any changes made to the page. 

PS: my first github pull-request, yeah! Still need to learn more JavaScript tho.