Add `&nomutepreview` option and apply `&volume` to previews
When streaming video, it was already possible to monitor input audio by
showing video controls for the local video preview and pressing the
unmute button.  This PR adds two features to enhance this functionality:

- Add an option, `&nomutepreview`, to automatically unmute the preview.

  If passed in sender mode, the preview still starts muted on the device
  selection screen, but gets unmuted when you press "start".  This is to
  avoid undesired results if the wrong audio device is preselected.

  The option can also be passed in director mode in order to monitor the
  director's own stream.  In this case it takes effect as soon as the
  director enables their microphone/video, since there's no separate
  device selection screen.

- Apply the `&volume` setting to local video previews (again, in both
  sender and director mode).

  By itself, this does not unmute the preview, but determines what
  volume you get if you press the unmute button.  This can also be
  combined with `&nomutepreview`.

Documentation is included, but I wasn't able to test the formatting (I
don't know how to replicate the Markdown-to-HTML process used for the
docs site).

Features _not_ included:

- Monitoring a subset of input audio devices

- Sending monitor output to a non-default audio device

Still, this is good enough for me.  My use case involves setting an
Elgato HDMI capture device as the video/audio input in order to stream
gameplay (directly through vdo.ninja, no OBS).  I need to monitor audio
in order to hear the game.